Refactor
* added LoginService class
* added TaskService class
* added UserService class
